requirements.txt
pyfiglet
python-vlc
keyboard